agent_ui: Remove `message_editor` module
